Add XML API controller and dynamic query builder service
Introduces XmlController for XML processing via API and adds DynamicQueryBuilder service for building SQL queries dynamically. Fixes file size string rounding to use MidpointRounding.ToEven, corrects float step parsing in BootstrapInputNumber, adjusts Table pagination calculation, and ensures proper file stream disposal in UploadFileExtensions. Cleans up attribute builder logic in DynamicObjectContext.

+++ src/BootstrapBlazor/Components/InputNumber/BootstrapInputNumber.razor.cs
@@ -245,7 +245,7 @@ public partial class BootstrapInputNumber<TValue>
                 val = (TValue)(object)(short)(@short + short.Parse(step));
                 break;
             case float @float:
-                val = (TValue)(object)(@float + float.Parse(step));
+                val = (TValue)(object)(@float + (float)double.Parse(step));
                 break;
             case double @double:
                 val = (TValue)(object)(@double + double.Parse(step));
